fix: ensure that only post's owner can delete it

+++ src/controllers/postsController.js
@@ -35,7 +35,12 @@ export const publishPosts = async (req, res) => {
 
 export const deletePosts = async (req, res) => {
   const { id: postId } = req.params;
+  const { userId } = res.locals;
   try {
+    const { rows: user } = await postsRepository.getPostUser(postId);
+    if (user[0]?.user_id !== userId) {
+      return res.status(401).send('You can only edit your own posts!');
+    }
     await metadatasRepository.deleteMetadataByPostId(postId);
     await hashtagsRepository.deleteHashtagsPostsByPostId(postId);
     await postsRepository.deletePostById(postId);
